+++ src/realm/array_packed.cpp
@@ -202,7 +202,7 @@ size_t ArrayPacked::parallel_subword_find(const Array& arr, size_t start, size_t
 
     unaligned_word_iter it((uint64_t*)arr.m_data, start * arr.m_width);
     uint64_t vector = 0;
-    while (total_bit_count_left >= (int64_t)bit_count_pr_iteration) {
+    while (total_bit_count_left >= (unsigned)bit_count_pr_iteration) {
         const auto word = it.get(bit_count_pr_iteration);
         vector = bitwidth_cmp<Cond>(MSBs, word, search_vector);
         if (vector) {
